fix(app): compile errors (Long/Int cmp, FlowRow OptIn) + CI releases & versioning
- EncryptedDataSource: compare length with C.LENGTH_UNSET.toLong()
- OptIn ExperimentalLayoutApi for FlowRow (settings, vault)
- CI: publish GitHub Release with versioned APK (v0.1.0-bN), inject build label/number

+++ app/src/main/kotlin/com/cripta/app/viewer/EncryptedDataSource.kt
@@ -26,7 +26,7 @@ class EncryptedDataSource(
         uri = dataSpec.uri
         transferInitializing(dataSpec)
         channel.position(dataSpec.position)
-        bytesRemaining = if (dataSpec.length != C.LENGTH_UNSET) {
+        bytesRemaining = if (dataSpec.length != C.LENGTH_UNSET.toLong()) {
             dataSpec.length
         } else {
             plaintextLength - dataSpec.position
